refactor: slimmed down the onboarding process, made ui more accessible
