+++ client/src/lib/components/home/home-page.tsx
@@ -5,6 +5,7 @@ import { motion } from "motion/react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import RecentBadges from "@/lib/components/achievements/recent-badges";
+import { AnimatedMochi as Mochi } from "@/lib/components/common/animated-mochi";
 import { MochiDefault, MochiHappy } from "@/lib/components/common/mochi";
 import DateScroll from "@/lib/components/home/date-scroll";
 import { useCheckinStore } from "@/lib/components/providers/zustand-provider";
@@ -120,7 +121,7 @@ export default function HomePage() {
 				<div className="relative flex flex-col gap-6 md:flex-row md:items-center md:justify-between">
 					<div className="flex justify-center md:justify-start">
 						<div className="relative">
-							<MochiHappy className="size-36 drop-shadow-md md:size-40 animate-float" />
+							<Mochi className="size-36 drop-shadow-md md:size-40" />
 							{/* Soft shadow beneath Mochi */}
 							<div
 								className="absolute -bottom-2 left-1/2 -translate-x-1/2 h-4 w-24 rounded-full bg-primary/8 blur-lg"
