+++ client/src/lib/components/onboarding/intermediate-step-2.tsx
@@ -8,7 +8,7 @@ export function IntermediateStepTwo() {
 	useEffect(() => {
 		const timer = setTimeout(() => {
 			goNext();
-		}, 2000);
+		}, 1200);
 
 		return () => clearTimeout(timer);
 	}, [goNext]);
