+++ client/src/lib/components/onboarding/onboarding-flow.tsx
@@ -49,7 +49,7 @@ export function OnboardingFlow() {
 				</div>
 
 				<div className="pb-8 flex justify-center gap-2 bg-background">
-					{[1, 2, 3, 4, 5, 6].map((i) => (
+					{[1, 2, 3, 4].map((i) => (
 						<motion.div
 							key={i}
 							initial={false}
